requirements.txt
easydict==1.9
numpy==1.21.6
opencv_python==4.5.5.64
Pillow==9.1.0
tensorboardX==2.5
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
